+++ src/components/App.js
@@ -18,8 +18,8 @@ const App = () => (
         <Route exact path="/login" component={Login} />
         <Route exact path="/sign-up" component={Signup} />
         <Route exact path="/verify-email" component={EmailVerification} />
-        <PrivateRoute path="/dashboard" component={Dashboard} />
-        {/* <Route path="/dashboard" component={Dashboard} /> */}
+        {/* <PrivateRoute path="/dashboard" component={Dashboard} /> */}
+        <Route path="/dashboard" component={Dashboard} />
         <Route component={NotFound} />
       </Switch>
     </Fragment>
